chore(xml): extract `Policy...` and `Rule...` out into separate module
